requirements.txt
aiogram>=2.11.2,<3.0
environs~=8.0.0
asyncpg
aioredis==1.3.1
asgiref
Pillow==8.0.1